requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.3
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
